+++ src/contour/ContourApp.h
@@ -36,6 +36,7 @@ class ContourApp: public crispy::App
     int terminfoAction();
     int configAction();
     int integrationAction();
+    int imageAction();
 };
 
 } // namespace contour